Fix large number overflow

+++ api/v1/client.go
@@ -961,7 +961,7 @@ func (c *storageClient) Delegations(ctx context.Context, r *http.Request) (*Dele
 			return nil, common.ErrStorageError
 		}
 
-		d.Amount = d.Shares * escrowBalanceActive / escrowTotalSharesActive
+		d.Amount = uint64(float64(d.Shares) * float64(escrowBalanceActive) / float64(escrowTotalSharesActive))
 
 		ds.Delegations = append(ds.Delegations, d)
 	}
@@ -1030,7 +1030,7 @@ func (c *storageClient) DebondingDelegations(ctx context.Context, r *http.Reques
 			)
 			return nil, common.ErrStorageError
 		}
-		d.Amount = d.Shares * escrowBalanceDebonding / escrowTotalSharesDebonding
+		d.Amount = uint64(float64(d.Shares) * float64(escrowBalanceDebonding) / float64(escrowTotalSharesDebonding))
 		ds.DebondingDelegations = append(ds.DebondingDelegations, d)
 	}
